backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
neo4j==5.20.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pytest==8.3.2
